requirements.txt
astroid==2.0
Django==2.0.7
django-filter==2.0.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
pylint==2.0.0
python-decouple==3.1
pytz==2018.5
six==1.11.0
wrapt==1.10.11
